Fixed problem with using signed chars
Should have been using unsigned

Also use long long for the filesize

+++ Axona/CPP/Utils.h
@@ -1,9 +1,9 @@
 #include <fstream>
 #include <vector>
 
-long GetFileSize(std::string filename)
+long long GetFileSize(std::string filename)
 {
-    long length;
+    long long length;
     std::ifstream in(filename, std::ifstream::ate | std::ifstream::binary);
     length = in.tellg();
     return length;
